+++ example/webpack.config.js
@@ -2,6 +2,10 @@ const {join:pathJoin} = require('path');
 const VueLoaderPlugin = require('vue-loader/lib/plugin');
 const HtmlWebpackPlugin = require("html-webpack-plugin");
 
+function resolve(dir) {
+    return pathJoin(__dirname, '..', dir)
+}
+  
 module.exports = {
     mode: 'development',
     entry:{
@@ -13,7 +17,11 @@ module.exports = {
     },
     resolve: {
         extensions: ['.js', '.vue'],
-        modules: ['./components', 'node_modules']        
+        modules: ['./components', 'node_modules'],
+        alias: {
+            'vue$': 'vue/dist/vue.esm.js',
+            '@': pathJoin(__dirname, './src')
+          }
       },
 
     module: {        
@@ -23,8 +31,27 @@ module.exports = {
                 use: ['vue-loader']
             },
             {
+                test: /\.js$/,
+                loader: 'babel-loader',
+                include: [pathJoin(__dirname, 'src'), pathJoin(__dirname, 'node_modules/element-ui')],
+                exclude: [pathJoin(__dirname, 'node_modules/element-ui/lib/element-ui.common.js')]
+                // include: [resolve('src'), resolve('test')]
+               },
+            {
                 test: /\.(scss|css)$/,             
                 loader: ['vue-style-loader', 'css-loader', 'sass-loader']
+            },
+            {
+                test: /\.less$/,             
+                loader: ['vue-style-loader', 'css-loader', 'less-loader']
+            },
+            {
+              test: /\.(png|jpe?g|gif|svg)(\?.*)?$/,
+              loader: 'url-loader'
+            },
+            {
+              test: /\.(woff2?|eot|ttf|otf)(\?.*)?$/,
+              loader: 'url-loader'
             }
         ]
     },