+++ src/helloWorld.1.vue
@@ -13,7 +13,7 @@ export default {
 </script>
 
 <!-- Add "scoped" attribute to limit CSS to this component only -->
-<style scoped>
+<style scoped lang="less">
 .hello {
   color: #42b983;
 }